move UDP-specific things in OSCPortTest into a new class, UDPTransportTest
